Djangotest/requirements.txt
Django==5.2.7
mysqlclient==2.2.7
sqlparse==0.5.3
asgiref==3.10.0
tzdata==2025.2
gunicorn==23.0.0
debugpy==1.8.17
djangorestframework==3.16.1